# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/deadlines": {
      "post": {
        "operationId": "DeadlinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeadlineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Échéance créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deadline"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une nouvelle échéance",
        "tags": [
          "deadlines"
        ]
      },
      "get": {
        "operationId": "DeadlinesController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "nouvelle",
                "en cours",
                "en attente",
                "complétée",
                "annulée"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "critique",
                "haute",
                "moyenne",
                "basse"
              ],
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "privée",
                "équipe",
                "département",
                "organisation"
              ],
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creatorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des échéances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deadline"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer toutes les échéances",
        "tags": [
          "deadlines"
        ]
      }
    },
    "/deadlines/{id}": {
      "get": {
        "operationId": "DeadlinesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Échéance trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deadline"
                }
              }
            }
          },
          "404": {
            "description": "Échéance non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une échéance par son ID",
        "tags": [
          "deadlines"
        ]
      },
      "patch": {
        "operationId": "DeadlinesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeadlineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Échéance mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deadline"
                }
              }
            }
          },
          "404": {
            "description": "Échéance non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une échéance",
        "tags": [
          "deadlines"
        ]
      },
      "delete": {
        "operationId": "DeadlinesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Échéance supprimée"
          },
          "404": {
            "description": "Échéance non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une échéance",
        "tags": [
          "deadlines"
        ]
      }
    },
    "/deadlines/project/{projectId}": {
      "get": {
        "operationId": "DeadlinesController_findByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID du projet",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des échéances par projet",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deadline"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les échéances par projet",
        "tags": [
          "deadlines"
        ]
      }
    },
    "/deadlines/user/{userId}": {
      "get": {
        "operationId": "DeadlinesController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des échéances par utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deadline"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les échéances par utilisateur",
        "tags": [
          "deadlines"
        ]
      }
    },
...
    "/comments/deadline/{deadlineId}": {
      "get": {
        "operationId": "CommentsController_findByDeadline",
        "parameters": [
          {
            "name": "deadlineId",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commentaires de l'échéance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les commentaires par échéance",
        "tags": [
          "comments"
        ]
      }
    },
...
    "/attachments/deadline/{deadlineId}": {
      "get": {
        "operationId": "AttachmentsController_findByDeadline",
        "parameters": [
          {
            "name": "deadlineId",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des pièces jointes de l'échéance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les pièces jointes par échéance",
        "tags": [
          "attachments"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestion d'Échéances",
    "description": "API pour l'application de gestion d'échéances du secteur défense",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDeadlineDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre de l'échéance",
            "example": "Réunion de projet"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée",
            "example": "Discuter des avancées du projet XYZ"
          },
          "deadlineDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date et heure de l'échéance",
            "example": "2025-05-15T14:00:00.000Z"
          },
          "creatorId": {
            "type": "string",
            "description": "ID de l'utilisateur créateur",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "priority": {
            "type": "string",
            "description": "Priorité de l'échéance",
            "enum": [
              "critique",
              "haute",
              "moyenne",
              "basse"
            ],
            "example": "moyenne"
          },
          "status": {
            "type": "string",
            "description": "Statut de l'échéance",
            "enum": [
              "nouvelle",
              "en cours",
              "en attente",
              "complétée",
              "annulée"
            ],
            "example": "nouvelle"
          },
          "visibility": {
            "type": "string",
            "description": "Visibilité de l'échéance",
            "enum": [
              "privée",
              "équipe",
              "département",
              "organisation"
            ],
            "example": "équipe"
          },
          "projectId": {
            "type": "string",
            "description": "ID du projet associé (optionnel)",
            "example": "7a9d0f35-d562-4219-c732-d35cf1a2e5f9"
          }
        },
        "required": [
          "title",
          "description",
          "deadlineDate",
          "creatorId",
          "priority",
          "status",
          "visibility"
        ]
      },
      "Deadline": {
        "type": "object",
        "properties": {}
      },
      "UpdateDeadlineDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Titre de l'échéance",
            "example": "Réunion de projet"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée",
            "example": "Discuter des avancées du projet XYZ"
          },
          "deadlineDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date et heure de l'échéance",
            "example": "2025-05-15T14:00:00.000Z"
          },
          "creatorId": {
            "type": "string",
            "description": "ID de l'utilisateur créateur",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "priority": {
            "type": "string",
            "description": "Priorité de l'échéance",
            "enum": [
              "critique",
              "haute",
              "moyenne",
              "basse"
            ],
            "example": "moyenne"
          },
          "status": {
            "type": "string",
            "description": "Statut de l'échéance",
            "enum": [
              "nouvelle",
              "en cours",
              "en attente",
              "complétée",
              "annulée"
            ],
            "example": "nouvelle"
          },
          "visibility": {
            "type": "string",
            "description": "Visibilité de l'échéance",
            "enum": [
              "privée",
              "équipe",
              "département",
              "organisation"
            ],
            "example": "équipe"
          },
          "projectId": {
            "type": "string",
            "description": "ID du projet associé (optionnel)",
            "example": "7a9d0f35-d562-4219-c732-d35cf1a2e5f9"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {}
      },
...
      "Attachment": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}